fixed syntax

+++ src/timeman.cpp
@@ -129,10 +129,9 @@ void TimeManagement::init(Search::LimitsType& limits,
 
         // Limit the maximum possible time for this move
         optimumTime = TimePoint(optScale * timeLeft);
-        maximumTime =
-          TimePoint(std::min(double(max_Time_Tune / 1000.00) * limits.time[us] - moveOverhead,
-                             maxScale * optimumTime))
-          - 10;
+        maximumTime = TimePoint(std::min((max_Time_Tune / 1000.00) * limits.time[us] - moveOverhead,
+                                         maxScale * optimumTime))
+                    - 10;
     }
 
     if (options["Ponder"])
